requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.0